Add files via upload
The projects reveals useful information about the bikes sales and the customers. After cleaning the data in excel, the relationship of count of purchase with each of the columns has been described perfectly by developing an interactive and visually pleasing dashboard in the final sheet. the dashboard uses slicers. this will help the usrer to gain information about a specific group of customers.